requirements.txt
# Web Framework & Server
fastapi==0.119.0
uvicorn==0.37.0

# Database & Migrations
SQLAlchemy==2.0.44
alembic==1.17.0
asyncpg==0.30.0
psycopg2-binary==2.9.11

# Redis (The fix for your error)
redis>=5.0.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.1.1
email-validator==2.3.0

# Pydantic & Utilities
pydantic==2.12.2
pydantic-settings==2.11.0